+++ .travis.yml
@@ -1,19 +1,14 @@
 language: python
-
 python:
   - "2.7"
   - "3.5"
   - "3.6"
-
 os:
   - linux
 #  - osx
-  
 sudo: false
-  
 env:
   - PATH=$HOME/miniconda/bin:$PATH LD_LIBRARY_PATH=$HOME/miniconda/lib C_INCLUDE_PATH=$HOME/miniconda/include
-
 before_install:
   - if [[ "$TRAVIS_PYTHON_VERSION" == "2.7" ]]; then
       wget https://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O miniconda.sh;
@@ -25,19 +20,19 @@ before_install:
   - conda config --add channels conda-forge
   - conda config --set always_yes yes --set changeps1 no
   - conda update conda
-  - conda install anaconda-client
-
+  # - conda install anaconda-client
 install:
-  - conda install python=$TRAVIS_PYTHON_VERSION nose six numpy h5py scipy scikit-image pywavelets mkl-devel mkl_fft python-coveralls dxchange numexpr
+  - "conda install python=$TRAVIS_PYTHON_VERSION nose six numpy h5py scipy
+    scikit-image pywavelets mkl-devel mkl_fft python-coveralls dxchange
+    numexpr"
   - if [[ "$TRAVIS_PYTHON_VERSION" == "2.7" ]]; then
       conda install futures;
     fi
   - conda info -a
   - conda list --explicit
-  - python setup.py build_ext --inplace
-
-script: 
+  - python build.py
+  - python setup.py install
+script:
   - nosetests test --with-coverage
-
 after_success:
   - coveralls
