+++ build.sh
@@ -1,2 +1,3 @@
 #!/bin/bash
+$PYTHON build.py
 $PYTHON setup.py install
