--- setup_nocompile.py
@@ -1,35 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-from setuptools import setup, find_packages
-
-setup(
-    name='tomopy',
-    packages=find_packages(exclude=['test*']),
-    version=open('VERSION').read().strip(),
-    include_package_data=True,
-    ext_modules=[],
-    zip_safe=False,
-    author='Doga Gursoy',
-    author_email='dgursoy@aps.anl.gov',
-    description='Tomographic Reconstruction in Python.',
-    keywords=['tomography', 'reconstruction', 'imaging'],
-    url='http://tomopy.readthedocs.org',
-    download_url='http://github.com/tomopy/tomopy.git',
-    license='BSD-3',
-    platforms='Any',
-    classifiers=[
-        'Development Status :: 4 - Beta',
-        'License :: OSI Approved :: BSD License',
-        'Intended Audience :: Science/Research',
-        'Intended Audience :: Education',
-        'Intended Audience :: Developers',
-        'Natural Language :: English',
-        'Operating System :: OS Independent',
-        'Programming Language :: Python :: 2',
-        'Programming Language :: Python :: 2.7',
-        'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.5',
-        'Programming Language :: Python :: 3.6',
-        'Programming Language :: C']
-)
